servers/biometric_signal_sensor_interface/requirements.txt
numpy
matplotlib
threading
ConfigParser
pandas
csv
ctypes
scipy
numbers
Flask
Jinja2
MarkupSafe
PyTango
SQLAlchemy
WSGIserver
Werkzeug
plotlyrequests
requests-oauthlib
urllib3